# app/Http/Resources/CategoryResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

/**
 *  @OA\Schema(
 *      schema="CategoryResource",
 *
 * @OA\Property(property="id", type="integer" , example=1),
 * @OA\Property(property="name", type="string" , example="example"),
 * @OA\Property(property="icon", type="string" , example="example"),
 *
 * )
 */

class CategoryResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     */
    public function toArray(Request $request): array
    {
        return [
            "id"=>$this->id,
            "name"=>$this->name,
            "icon"=>$this->icon_url
        ];

    }
}