# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API documentation for the library project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a paginated list of all books",
                "operationId": "921c595372ab399821e267fcf05cbaa1",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "create a  book (Admin ONLY)",
                "operationId": "7ab92284110ace5d0f7de52f372e70c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a single book by id",
                "operationId": "e079e587c1fee53fe150984526abf3ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "update a book (Admin only)",
                "operationId": "b47b1606570ba86a105bf62f04de6b7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successful (admin only)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "delete a book (admin only)",
                "operationId": "95c7fc7d78f9dc3884a5bb898b7ead11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "book deleted successful (admin only)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "get all category",
                "operationId": "b1b247b2b491b0491c3c728dc33f23f0",
                "responses": {
                    "200": {
                        "description": "successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "create a category (admin only)",
                "operationId": "c86cd7a85ad06a4fe905bfab62146952",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "category created successfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "get category by id",
                "operationId": "486849467f411ec5e4bdf111f50ede87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "update the category (admin only)",
                "operationId": "b648463d94e0caf7b8d842a1008789b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "delete a category (admin only)",
                "operationId": "491dfe5b6b58b480181690ede6550e44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "category deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreBookRequest": {
                "required": [
                    "ISBN",
                    "title",
                    "description",
                    "published_date",
                    "category_id",
                    "author_id"
                ],
                "properties": {
                    "ISBN": {
                        "type": "string",
                        "example": "9781234567897"
                    },
                    "title": {
                        "type": "string",
                        "example": "Clean Architecture"
                    },
                    "description": {
                        "type": "string",
                        "example": "A book about clean architecture"
                    },
                    "published_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book_upload": {
                        "description": "PDF File",
                        "type": "string",
                        "format": "binary"
                    },
                    "image": {
                        "description": "Cover image",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "UpdateBookRequest": {
                "properties": {
                    "ISBN": {
                        "type": "string",
                        "example": "9781234567897"
                    },
                    "title": {
                        "type": "string",
                        "example": "Clean Architecture"
                    },
                    "description": {
                        "type": "string",
                        "example": "A book about clean architecture"
                    },
                    "published_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book_upload": {
                        "description": "PDF File",
                        "type": "string",
                        "format": "binary"
                    },
                    "image": {
                        "description": "Cover image",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "StoreCategoryRequest": {
                "required": [
                    "name",
                    "icon"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "icon": {
                        "description": "icon image",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "UpdateCategoryRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "icon": {
                        "description": "icon image",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "BookResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ISBN": {
                        "type": "string",
                        "example": "9781234567897"
                    },
                    "title": {
                        "type": "string",
                        "example": "Clean Architecture"
                    },
                    "description": {
                        "type": "string",
                        "example": "A book about clean architecture"
                    },
                    "published_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "example"
                    },
                    "icon": {
                        "type": "string",
                        "example": "example"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}